Extended functions merge into Develop

+++ src/controllers/user.js
@@ -3,54 +3,43 @@ const uuid = require('uuid');
 const models = require("../../models");
 
 const createUser = async (req, res) => {
-    const name = req.body.name;
-    const surname = req.body.surname;
-    const email = req.body.email;
-    const username = req.body.username;
+  try {
+    const { name, surname, email } = req.body;
+    const username = helpers.hasNoSpaces(req.body.username);
     const password = await helpers.encryptPassword(req.body.password);
 
-    try {
-    const post = await models.User.create({
+    if (username) {
+      const user = await models.User.create({
         Usr_name: name,
         Usr_surname: surname,
         Usr_email: email,
         Usr_username: username,
         Usr_password: password
-    });
+      });
 
-    const wallet = await models.Wallet.create({
+      const wallet = await models.Wallet.create({
         Wal_id: uuid.v4(),
-        Usr_id: post.Usr_id,
+        Usr_id: user.Usr_id,
         Wtyp_id: 1,
         Wal_balance: 0.00,
         Wal_state: "Active"
-    });
+      });
 
-    return res.status(201).json({
-      post,
-    });
+      return res.status(201).json({
+        user: user,
+        wallet: wallet
+      });
+    }
+    return res.status(400).send("Username can't contain spaces");
   } catch (error) {
-    return res.status(500).json({error: error.message})
+    return res.status(500).json({ error: error.message })
   }
 }
 
 const getAllUsers = async (req, res) => {
   try {
-    const posts = await models.User.findAll({
-      include: [
-        {
-          model: models.Wallet,
-          as: 'possess',
-        include: [
-          {
-           model: models.Movement,
-           as: 'modifies',
-          }
-         ]
-        },
-      ]
-    });
-    return res.status(200).json({ posts });
+    const users = await models.User.findAll();
+    return res.status(200).json({ users: users });
   } catch (error) {
     return res.status(500).send(error.message);
   }
@@ -59,23 +48,11 @@ const getAllUsers = async (req, res) => {
 const getUserByUsername = async (req, res) => {
   try {
     const { username } = req.params;
-    const post = await models.User.findOne({
-      where: { Usr_username: username },
-      include: [
-        {
-          model: models.Wallet,
-          as: 'possess',
-          include: [
-           {
-            model: models.Movement,
-            as: 'modifies',
-           }
-          ]
-        }
-      ]
+    const user = await models.User.findOne({
+      where: { Usr_username: username }
     });
-    if (post) {
-      return res.status(200).json({ post });
+    if (user) {
+      return res.status(200).json({ user: user });
     }
     return res.status(404).send('User with the specified username does not exists');
   } catch (error) {
@@ -86,26 +63,27 @@ const getUserByUsername = async (req, res) => {
 const updateUser = async (req, res) => {
   try {
     const { username } = req.params;
-    const name = req.body.name;
-    const surname = req.body.surname;
-    const usrname = req.body.username;
-    const email = req.body.email;
+    const usrname = helpers.hasNoSpaces(req.body.username);
+    const { name, surname, email } = req.body;
     const password = await helpers.encryptPassword(req.body.password);
 
-    const [ updated ] = await models.User.update({
+    if (usrname) {
+      const [updated] = await models.User.update({
         Usr_name: name,
         Usr_surname: surname,
         Usr_email: email,
         Usr_username: usrname,
         Usr_password: password
-    }, {
-      where: { Usr_username: username }
-    });
-    if (updated) {
-      const updatedUser = await models.User.findOne({ where: { Usr_username: usrname } });
-      return res.status(200).json({ user: updatedUser });
+      }, {
+        where: { Usr_username: username }
+      });
+      if (updated) {
+        const updatedUser = await models.User.findOne({ where: { Usr_username: usrname } });
+        return res.status(200).json({ user: updatedUser });
+      }
+      throw new Error('User not updated');
     }
-    throw new Error('User not found');
+    return res.status(400).send("Username can't contain spaces");
   } catch (error) {
     return res.status(500).send(error.message);
   }
@@ -114,44 +92,57 @@ const updateUser = async (req, res) => {
 const deleteUser = async (req, res) => {
   try {
     const { username } = req.params;
-    const deleted = await models.User.destroy({
-      where: { Usr_username: username }
-    });
-    if (deleted) {
-      return res.status(200).send("User deleted");
+    const findUser = await models.User.findOne({ where: { Usr_username: username } });
+
+    if (findUser) {
+      const wallets = await models.Wallet.findAll({ where: { Usr_id: findUser.Usr_id } });
+      
+      if (wallets.length > 0) {
+        return res.status(400).send("This User has associated Wallets. Please delete them first");
+      }
+      const deleted = await models.User.destroy({
+        where: { Usr_username: username }
+      });
+      if (deleted) {
+        return res.status(200).send("User " + username + " deleted");
+      }
     }
-    throw new Error("User not found");
-  } catch (error) {
-    return res.status(500).send(error.message);
-  }
+    return res.status(404).send("Specified User not found");
+} catch (error) {
+  return res.status(500).send(error.message);
+}
 };
 
 const validateUser = async (req, res) => {
   try {
     const { username, password } = req.body;
-    const post = await models.User.findOne({
+    const user = await models.User.findOne({
       where: { Usr_username: username },
       include: [
         {
           model: models.Wallet,
           as: 'possess',
           include: [
-           {
-            model: models.Movement,
-            as: 'modifies',
-           }
+            {
+              model: models.Movement,
+              as: 'modifies_sender',
+            },
+            {
+              model: models.Movement,
+              as: 'modifies_recipient', 
+            }
           ]
         }
       ]
     });
-    if (post) {
-      const val = await helpers.matchPassword(password,post.Usr_password);
-      if(val){
-        return res.status(200).json({ post }); //Si se autenticó correctamente, le devuelve el user con su wallet
+    if (user) {
+      const val = await helpers.matchPassword(password, user.Usr_password);
+      if (val) {
+        return res.status(200).json({ user: user }); //Si se autenticó correctamente, le devuelve el user con su wallet
       } else {
         return res.status(401).send('The password is incorrect. Please try again');
       }
-      
+
     }
     return res.status(404).send('User with specified username does not exists');
   } catch (error) {
